[SR] Avoid boxing inputs in DictConstruct/ListUnpack
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74250

The `DictConstruct`/`ListUnpack` implementations currently put all of their inputs onto a stack before calling the JIT implementation in `vararg_functions.cpp`. This was done to avoid code duplication, but it's quite wasteful since it causes extra heap allocations and, potentially, refcount bumps.

Given that these two ops are quite common and the code duplication is only a few lines, it seems reasonable to avoid this cost.
ghstack-source-id: 151897634

Differential Revision: [D34901245](https://our.internmc.facebook.com/intern/diff/D34901245/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D34901245/)!