[quant][fx] Merge is_general_tensor_shape_op into is_general_tensor_value_op in QuantizeHandler
Summary:
Currently the behavior for general tensor shape op and general tensor value op are the same, so we can remove
this flag and merge with the is_general_tensor_value_op flag.

is_general_tensor_value_op flag is used in two places in prepare:
(1). dtype propgation: we only do dtype propgation when this flag is true (this will be refactor in the future to be more systematic)
(2). observer sharing, we'll use the input observer instance as output observer for an op if this flag is True

Test Plan:
python test/test_quantization.py TestQuantizeFx
python test/test_quantization.py TestQuantizeFxOps

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 93c9660d36239e612ebff8f7a07ac3ea47aad325
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74601